+++ packages/@ourworldindata/types/src/index.ts
@@ -1,1 +1,8 @@
 export { type FormattingOptions, SubNavId } from "./Wordpress.js"
+export {
+    type DonateSessionResponse,
+    type DonationCurrencyCode,
+    type DonationInterval,
+    type DonationRequest,
+    DonationRequestTypeObject,
+} from "./DonationTypes.js"
