+++ packages/@ourworldindata/utils/src/dbTypes/PostTags.ts
@@ -1,0 +1,7 @@
+/** the entity in the `post_tags` table */
+export interface PostTag {
+    post_id: number
+    tag_id: number
+    createdAt: Date
+    updatedAt: Date | null
+}
