+++ packages/@ourworldindata/utils/src/dbTypes/Tags.ts
@@ -1,0 +1,11 @@
+/** the entity in the `tags` table */
+export interface Tag {
+    id: number
+    name: string
+    createdAt: Date
+    updatedAt: Date
+    parentId: number
+    isBulkImport: boolean
+    specialType: string
+    slug: string | null
+}
