+++ packages/@ourworldindata/utils/src/domainTypes/ContentGraph.ts
@@ -1,0 +1,71 @@
+import { TopicId } from "./Various.js"
+
+export interface EntryMeta {
+    slug: string
+    title: string
+    excerpt: string
+    kpi: string
+}
+
+export interface CategoryWithEntries {
+    name: string
+    slug: string
+    entries: EntryMeta[]
+    subcategories: CategoryWithEntries[]
+}
+export enum GraphDocumentType {
+    Topic = "topic",
+    Article = "article",
+}
+
+export interface AlgoliaRecord {
+    id: number
+    title: string
+    type: GraphType | GraphDocumentType
+    image?: string
+}
+
+export interface DocumentNode {
+    id: number
+    title: string
+    slug: string
+    content: string | null // if content is empty
+    type: GraphDocumentType
+    image: string | null
+    parentTopics: Array<TopicId>
+}
+
+export interface CategoryNode {
+    name: string
+    slug: string
+    pages: any
+    children: any
+}
+
+export enum GraphType {
+    Document = "document",
+    Chart = "chart",
+}
+
+export interface ChartRecord {
+    id: number
+    title: string
+    slug: string
+    type: GraphType.Chart
+    parentTopics: Array<TopicId>
+}
+
+export interface PostReference {
+    id: number | string
+    title: string
+    slug: string
+}
+
+export interface EntryNode {
+    slug: string
+    title: string
+    // in some edge cases (entry alone in a subcategory), WPGraphQL returns
+    // null instead of an empty string)
+    excerpt: string | null
+    kpi: string
+}
