+++ packages/@ourworldindata/utils/src/domainTypes/DataValues.ts
@@ -1,0 +1,23 @@
+export interface DataValueQueryArgs {
+    variableId?: number
+    entityId?: number
+    chartId?: number
+    year?: number
+}
+
+export interface DataValueConfiguration {
+    queryArgs: DataValueQueryArgs
+    template: string
+}
+
+export interface DataValueResult {
+    value: number
+    year: number
+    unit?: string
+    entityName: string
+}
+
+export interface DataValueProps extends DataValueResult {
+    formattedValue?: string
+    template: string
+}
