+++ packages/@ourworldindata/utils/src/domainTypes/DeployStatus.ts
@@ -1,0 +1,18 @@
+export enum DeployStatus {
+    queued = "queued",
+    pending = "pending",
+    // done = "done"
+}
+
+export interface DeployChange {
+    timeISOString?: string
+    authorName?: string
+    authorEmail?: string
+    message?: string
+    slug?: string
+}
+
+export interface Deploy {
+    status: DeployStatus
+    changes: DeployChange[]
+}
