+++ packages/@ourworldindata/utils/src/domainTypes/Toc.ts
@@ -1,0 +1,11 @@
+export interface TocHeading {
+    text: string
+    html?: string // used by SectionHeading toc. Excluded from LongFormPage toc.
+    slug: string
+    isSubheading: boolean
+}
+
+export interface TocHeadingWithTitleSupertitle extends TocHeading {
+    title: string
+    supertitle?: string
+}
