+++ packages/@ourworldindata/utils/src/grapherTypes/GrapherConstants.ts
@@ -1,0 +1,5 @@
+// todo: remove when we ditch Year and YearIsDay
+export const EPOCH_DATE = "2020-01-21"
+
+/** The "plot" is a chart without any header, footer or controls */
+export const IDEAL_PLOT_ASPECT_RATIO = 1.8
