+++ packages/@ourworldindata/utils/src/index.ts
@@ -3,135 +3,119 @@ export type {
     OwidVariableDataTableConfigInteface,
     OwidChartDimensionInterface,
 } from "./OwidVariableDisplayConfigInterface.js"
-
 export {
-    type AlgoliaRecord,
-    type Annotation,
-    AxisAlign,
-    type BasicChartInformation,
-    BLOCK_WRAPPER_DATATYPE,
-    BlockImageSize,
-    checkIsBlockImageSize,
-    type BlockPositionChoice,
-    type Box,
-    type CategoryNode,
-    type CategoryWithEntries,
-    type ChartPositionChoice,
-    type ChartRecord,
-    ChartControlKeyword,
-    ChartTabKeyword,
-    type Color,
-    type ColumnSlug,
+    type GitCommit,
+    type Integer,
+    JsonError,
+    type SerializedGridProgram,
+    SiteFooterContext,
+    SuggestedChartRevisionStatus,
+    TaggableType,
+    type TopicId,
+    type OwidVariableId,
+    type RawPageview,
+} from "./domainTypes/Various.js"
+export {
     type DataValueConfiguration,
     type DataValueProps,
     type DataValueQueryArgs,
     type DataValueResult,
+} from "./domainTypes/DataValues.js"
+export {
+    type BreadcrumbItem,
+    type KeyInsight,
+    type KeyValueProps,
+} from "./domainTypes/Site.js"
+export {
+    type FormattedPost,
+    type IndexPost,
+    type FullPost,
+} from "./domainTypes/Posts.js"
+
+export {
+    type TocHeading,
+    type TocHeadingWithTitleSupertitle,
+} from "./domainTypes/Toc.js"
+
+export {
     type Deploy,
     type DeployChange,
     DeployStatus,
-    type DonateSessionResponse,
-    type DonationCurrencyCode,
-    type DonationInterval,
-    type DonationRequest,
-    DonationRequestTypeObject,
-    type EnrichedDetail,
-    type EnrichedFaq,
-    type FaqEntryData,
-    type DetailDictionary,
-    DimensionProperty,
-    type DocumentNode,
-    type EnrichedBlockAllCharts,
-    type EnrichedBlockAdditionalCharts,
-    type EnrichedBlockAside,
-    type EnrichedBlockBlockquote,
-    type EnrichedBlockCallout,
-    type EnrichedBlockChart,
-    type EnrichedBlockChartStory,
-    type EnrichedBlockExpandableParagraph,
-    type EnrichedBlockGraySection,
-    type EnrichedBlockHeading,
-    type EnrichedBlockHorizontalRule,
-    type EnrichedBlockHtml,
-    type EnrichedBlockImage,
-    type EnrichedBlockVideo,
-    type EnrichedBlockKeyInsights,
-    type EnrichedBlockKeyInsightsSlide,
-    type EnrichedBlockList,
-    type EnrichedBlockMissingData,
-    type EnrichedBlockNumberedList,
-    type EnrichedBlockProminentLink,
-    type EnrichedBlockPullQuote,
-    type EnrichedBlockRecirc,
-    type EnrichedBlockResearchAndWriting,
-    type EnrichedBlockResearchAndWritingLink,
-    type EnrichedBlockResearchAndWritingRow,
-    type EnrichedBlockScroller,
-    type EnrichedBlockSDGGrid,
-    type EnrichedBlockSDGToc,
-    type EnrichedBlockSideBySideContainer,
-    type EnrichedBlockSimpleText,
-    type EnrichedBlockStickyLeftContainer,
-    type EnrichedBlockStickyRightContainer,
-    type EnrichedBlockText,
-    type EnrichedTopicPageIntroRelatedTopic,
-    type EnrichedTopicPageIntroDownloadButton,
-    type EnrichedBlockTopicPageIntro,
-    type EnrichedChartStoryItem,
-    type EnrichedRecircLink,
-    type EnrichedScrollerItem,
-    type EnrichedSDGGridItem,
-    type EnrichedBlockEntrySummary,
-    type EnrichedBlockEntrySummaryItem,
-    type EntryMeta,
-    type EntryNode,
-    type EnrichedBlockTable,
-    type EnrichedBlockTableRow,
-    type EnrichedBlockTableCell,
-    EPOCH_DATE,
-    type FilterFnPostRestApi,
-    type FormattedPost,
-    type FullPost,
-    GdocsContentSource,
-    type GitCommit,
-    GraphDocumentType,
-    GraphType,
-    type GridParameters,
-    HorizontalAlign,
+} from "./domainTypes/DeployStatus.js"
+export { type Tag } from "./dbTypes/Tags.js"
+
+export { type ChartTag, type ChartTagJoin } from "./dbTypes/ChartTags.js"
+
+export {
     IDEAL_PLOT_ASPECT_RATIO,
-    type IndexPost,
-    type Integer,
-    JsonError,
+    EPOCH_DATE,
+} from "./grapherTypes/GrapherConstants.js"
+
+export {
+    type Annotation,
+    type Box,
+    type BasicChartInformation,
+    SortBy,
+    type SortConfig,
+    SortOrder,
+    type ValueRange,
+    type Year,
+    TimeBoundValue,
+    type TimeRange,
+    type Color,
+    type ColumnSlug,
     KeyChartLevel,
-    type KeyInsight,
-    type KeyValueProps,
-    type LinkedChart,
-    OwidGdocPublicationContext,
-    type OwidGdocErrorMessageProperty,
-    type OwidGdocErrorMessage,
-    OwidGdocErrorMessageType,
-    type OwidGdocLinkJSON,
-    type OwidGdocBaseInterface,
-    type OwidGdocPostContent,
-    type OwidGdocPostInterface,
-    type OwidGdocDataInsightContent,
-    type OwidGdocDataInsightInterface,
-    type MinimalDataInsightInterface,
-    DATA_INSIGHTS_INDEX_PAGE_SIZE,
-    type OwidGdoc,
-    type Tag,
-    OwidGdocType,
-    type OwidGdocStickyNavItem,
-    type OwidGdocJSON,
-    type OwidEnrichedGdocBlock,
-    type OwidRawGdocBlock,
-    type OwidArticleBackportingStatistics,
-    type OwidVariableId,
-    type ParseError,
+    type PrimitiveType,
+    DimensionProperty,
+    type RelatedChart,
+    ToleranceStrategy,
+    ScaleType,
+} from "./grapherTypes/GrapherTypes.js"
+
+export {
     Position,
     type PositionMap,
+    AxisAlign,
+    VerticalAlign,
+    type GridParameters,
+    HorizontalAlign,
+} from "./domainTypes/Layout.js"
+
+export {
+    type EntryMeta,
+    type EntryNode,
     type PostReference,
+    type CategoryNode,
+    type DocumentNode,
+    type CategoryWithEntries,
+    GraphDocumentType,
+    GraphType,
+    type AlgoliaRecord,
+    type ChartRecord,
+} from "./domainTypes/ContentGraph.js"
+export {
+    WP_BlockClass,
+    WP_BlockType,
+    WP_ColumnStyle,
+    WP_PostType,
     type PostRestApi,
-    type PrimitiveType,
+    type FilterFnPostRestApi,
+} from "./wordpress/WordpressTypes.js"
+
+export {
+    type Ref,
+    type RefDictionary,
+    type BlockPositionChoice,
+    type ChartPositionChoice,
+    type OwidEnrichedGdocBlock,
+    type OwidRawGdocBlock,
+    ChartControlKeyword,
+    ChartTabKeyword,
+    type EnrichedBlockAlign,
+    type RawBlockAlign,
+    type ParseError,
+    BlockImageSize,
+    checkIsBlockImageSize,
     type RawBlockAllCharts,
     type RawBlockAdditionalCharts,
     type RawBlockAside,
@@ -175,20 +159,99 @@ export {
     type RawBlockTableCell,
     type RawChartStoryValue,
     type RawRecircLink,
-    type RawDetail,
     type RawSDGGridItem,
     type RawBlockEntrySummary,
     type RawBlockEntrySummaryItem,
-    type RelatedChart,
-    type Ref,
-    type RefDictionary,
-    ScaleType,
-    type SerializedGridProgram,
+    type EnrichedBlockAllCharts,
+    type EnrichedBlockAdditionalCharts,
+    type EnrichedBlockAside,
+    type EnrichedBlockBlockquote,
+    type EnrichedBlockCallout,
+    type EnrichedBlockChart,
+    type EnrichedBlockChartStory,
+    type EnrichedBlockExpandableParagraph,
+    type EnrichedBlockGraySection,
+    type EnrichedBlockHeading,
+    type EnrichedBlockHorizontalRule,
+    type EnrichedBlockHtml,
+    type EnrichedBlockImage,
+    type EnrichedBlockVideo,
+    type EnrichedBlockKeyInsights,
+    type EnrichedBlockKeyInsightsSlide,
+    type EnrichedBlockList,
+    type EnrichedBlockMissingData,
+    type EnrichedBlockNumberedList,
+    type EnrichedBlockProminentLink,
+    type EnrichedBlockPullQuote,
+    type EnrichedBlockRecirc,
+    type EnrichedBlockResearchAndWriting,
+    type EnrichedBlockResearchAndWritingLink,
+    type EnrichedBlockResearchAndWritingRow,
+    type EnrichedBlockScroller,
+    type EnrichedBlockSDGGrid,
+    type EnrichedBlockSDGToc,
+    type EnrichedBlockSideBySideContainer,
+    type EnrichedBlockSimpleText,
+    type EnrichedBlockStickyLeftContainer,
+    type EnrichedBlockStickyRightContainer,
+    type EnrichedBlockText,
+    type EnrichedTopicPageIntroRelatedTopic,
+    type EnrichedTopicPageIntroDownloadButton,
+    type EnrichedBlockTopicPageIntro,
+    type EnrichedChartStoryItem,
+    type EnrichedRecircLink,
+    type EnrichedScrollerItem,
+    type EnrichedSDGGridItem,
+    type EnrichedBlockEntrySummary,
+    type EnrichedBlockEntrySummaryItem,
+    type EnrichedBlockTable,
+    type EnrichedBlockTableRow,
+    type EnrichedBlockTableCell,
+    type RawBlockResearchAndWritingRow,
+} from "./gdocTypes/ArchieMlComponents.js"
+export {
+    OwidGdocPublicationContext,
+    type OwidGdocErrorMessageProperty,
+    type OwidGdocErrorMessage,
+    OwidGdocErrorMessageType,
+    type OwidGdocLinkJSON,
+    type OwidGdocBaseInterface,
+    type OwidGdocPostContent,
+    type OwidGdocPostInterface,
+    type OwidGdocDataInsightContent,
+    type OwidGdocDataInsightInterface,
+    type MinimalDataInsightInterface,
+    DATA_INSIGHTS_INDEX_PAGE_SIZE,
+    type OwidGdoc,
+    type RawDetail,
+    OwidGdocType,
+    type OwidGdocStickyNavItem,
+    type OwidGdocJSON,
+    type FaqDictionary,
+    type EnrichedDetail,
+    type EnrichedFaq,
+    type DetailDictionary,
+    GdocsContentSource,
+    type OwidArticleBackportingStatistics,
+    type LinkedChart,
     DYNAMIC_COLLECTION_PAGE_CONTAINER_ID,
-    SiteFooterContext,
-    SortBy,
-    type SortConfig,
-    SortOrder,
+} from "./gdocTypes/Gdoc.js"
+
+export {
+    DataPageJsonTypeObject,
+    type DataPageJson,
+    type DataPageParseError,
+    type DataPageV2ContentFields,
+    type DataPageDataV2,
+    type DataPageRelatedData,
+    type DataPageRelatedResearch,
+    type PrimaryTopic,
+    type FaqLink,
+    type FaqEntryData,
+    type DisplaySource,
+} from "./gdocTypes/Datapage.js"
+
+export {
     type Span,
     type SpanBold,
     type SpanDod,
@@ -202,48 +265,8 @@ export {
     type SpanSubscript,
     type SpanSuperscript,
     type SpanUnderline,
-    SuggestedChartRevisionStatus,
-    type ChartTag,
-    type ChartTagJoin,
-    TaggableType,
-    type Time,
-    type TimeBound,
-    type TimeBounds,
-    TimeBoundValue,
-    type TimeRange,
-    type TocHeading,
-    type TocHeadingWithTitleSupertitle,
-    type TopicId,
     type UnformattedSpan,
-    type ValueRange,
-    VerticalAlign,
-    WP_BlockClass,
-    WP_BlockType,
-    WP_ColumnStyle,
-    WP_PostType,
-    type Year,
-    type PostRowWithGdocPublishStatus,
-    ToleranceStrategy,
-    type RawPageview,
-    type DataPageGdocContent,
-    DataPageJsonTypeObject,
-    type DataPageJson,
-    type DataPageParseError,
-    AllowedDataPageGdocFields,
-    type DataPageContentFields,
-    type DataPageV2ContentFields,
-    type DataPageDataV2,
-    type DataPageRelatedData,
-    type DataPageRelatedResearch,
-    type PrimaryTopic,
-    type FaqLink,
-    type FaqDictionary,
-    type RawBlockResearchAndWritingRow,
-    type EnrichedBlockAlign,
-    type RawBlockAlign,
-    type BreadcrumbItem,
-    type DisplaySource,
-} from "./owidTypes.js"
+} from "./gdocTypes/Spans.js"
 
 export {
     parsePostFormattingOptions,
@@ -259,6 +282,7 @@ export {
     type PostRowUnparsedFields,
     type PostRowEnriched,
     type PostRowRaw,
+    type PostRowWithGdocPublishStatus,
 } from "./dbTypes/Posts.js"
 
 export {
