+++ packages/@ourworldindata/utils/src/wordpress/WordpressTypes.ts
@@ -1,0 +1,68 @@
+export enum WP_PostType {
+    Post = "post",
+    Page = "page",
+}
+
+export interface PostRestApi {
+    slug: string
+    meta: {
+        owid_publication_context_meta_field?: {
+            immediate_newsletter?: boolean
+            homepage?: boolean
+            latest?: boolean
+        }
+    }
+    id: number
+    date: string
+    date_gmt: string
+    guid: {
+        rendered: string
+    }
+    modified: string
+    modified_gmt: string
+
+    status: string
+    type: WP_PostType
+    link: string
+    title: {
+        rendered: string
+    }
+    content: {
+        rendered: string
+        protected: boolean
+    }
+    excerpt: {
+        rendered: string
+        protected: boolean
+    }
+    author: number
+    featured_media: number
+    comment_status: string
+    ping_status: string
+    sticky: boolean
+    template: string
+    format: string
+    categories: number[]
+    tags: any[]
+    authors_name: string[]
+    featured_media_paths: {
+        thumbnail: string
+        medium_large: string
+    }
+}
+
+export type FilterFnPostRestApi = (post: PostRestApi) => boolean
+
+export enum WP_ColumnStyle {
+    StickyRight = "sticky-right",
+    StickyLeft = "sticky-left",
+    SideBySide = "side-by-side",
+}
+
+export enum WP_BlockClass {
+    FullContentWidth = "wp-block-full-content-width", // not an actual WP block yet
+}
+
+export enum WP_BlockType {
+    AllCharts = "all-charts",
+}
